Update README.md
This will be an app that people will be able to use to help save for trips. I've noticed in my own personal life that going on trips can sometimes be a little hard saving for, especially if the dicipline isn't there. What I would like for this app to do is to have all of the basic functions of a traveling app, but add an automated saving plan that will take monthly payments out of someones actual account after there real time expenses have been paid. Payments can come out daily, weekly or monthly. This will take the hassel out of trying to create the dicipline to save for trips, when it can be done for you.